I don't understand half of what's going on, but it barely works
Also there's a ton of modifications and a TON of undefined behaviour
because the GB doesn't initialize the memory correctly. As a result
everytime I turn on the ROM a lot of things are random that I didn't
setted to 0 yet, so it's quite hard to understand what's going on
because of that randomization. Someone should everything to 0 as soon
as possible (or the things that can be set to 0, at least).

Also, hardware.inc was missing a few declarations for the different
Tile Blocks (from $8000 up to $9000).