More progress to debugger wrapper
This time we changed the fb_in() and fb_out() names to fb_master() and
fb_slave() to better convey the underlying pseudo terminals.

We also redirected the STDERR to the pseudo terminal, because the SameBoy uses
it to output its first (and only) message: "SameBoy vX.XX.X". We need
that message in order to know *when* to SIGINT the process and bring
up the debugger, because if we do it earlier than that message we may
kill the SameBoy's process because it may not have completely set up
its signal handlers.

We still haven't changed the Pipe() naming system, though. It should
be changed somewhere in the future.

We probably need to refactor the inner loops in such a way that it
does show the internal logic. Right now it just a mess of if()s and
loops and variable declarations that will probably only get bigger as
we write more of the debugging logic and the ncurses' logic.