Improving the main Binary Search algorithm in the core 2D engine
This binary search algorithm is used to manage the tiles in the
VRAM. We probably will do this bookkepping outside of VBlank and write
the few necessary changes that need to made in RAM such so we don't do
way too much work while in VBlank.

The virtual array that this algotihm manages always has (2^n)-1
elements in it. Here's a description of how this array should behave,
because that affects the low-level assembly optimizations made.

array = {0, 1 ,2} where each index has a item

After expanding, the array contain the indexes
array = {0, 1, 2, 3, 4, 5, 6}

Where indexes 1, 3 and 5 are the previous indexes 0, 1 and 2
respectively. The indexes 0, 2, 4 and 6 are new indexes.

After expanding again, the array contain the indexes
array = {0 through 14}
Where each index of the previous array named I is now the (2*I + 1)
index. So 0 becomes 1, 1 becomes 3, 2 becomes 5 and so forth...

As you see, at each expanding level the array contain (2^level)-1
items. Also, the array expands around the central item. So, given a
255 array, you can simulate this "level-0" array using the indexes 0,
1 and 2 mapped into indexes 63, 127 and 191. After expanding, its
indexes 0, 1, 2, 3, 4, 5, and 6 are mapped into indexes 31, 63, 95,
127, 159 and 191. So you could say:

array_0 = {0 -> 63, 1 -> 127, 2 -> 191}
array_1 = {0 -> 31, 1 ->  63, 2 ->  95, 3 -> 127, 4 -> 159, 5 -> 191}

So you could say

array_0 -> array_1 = {0 -> 1, 1 -> 3, 2-> 5}

And you you see that the array_0[0] == array[1] and they ocuppy the
same spot in memory in relation to a "real" 255 array where they are
mapping into. So as this virtual array grows, there's no need to
shuffle data around because it expands around it's midpoint while
keeping the original virtual array sorted as it expands.

The fact that it keeps its items sorted as it expands makes this a
nice array for doing binary search on, as you can insert and remove
items without the need to sort this array or to copy its items into a
new bigger array to acomodate a new item (provided you always insert
items on it in a sorted manner). The downside is that you need a
complexier algorithm on the assembly level do deal with the quirkness
of this array, and it may grow unbalanced and need to sort quite early
depending on your usage.